Add DB role
