+++ src/backend/Database/Ts.TelemetryDatabase.Sql/Security/create_user.sql
@@ -1,0 +1,6 @@
+﻿-- TODO: Add user test/dev/prod. How to handle this?
+
+--USE [{DATABASE_NAME}]
+--CREATE USER [{IDENTITY}] FROM EXTERNAL PROVIDER;
+--EXEC sp_addrolemember 'db_datareader', [{IDENTITY}];
+--ALTER ROLE [db_execute_procedure_role] ADD MEMBER [{IDENTITY}]
