+++ src/backend/Database/Ts.TelemetryDatabase.Sql/Security/db_execute_procedure_role.sql
@@ -1,0 +1,4 @@
+﻿CREATE ROLE [db_execute_procedure_role]
+    AUTHORIZATION [dbo];
+
+GO
