+++ src/backend/Database/Ts.TelemetryDatabase.Sql/Stored Procedures/Register/UpsertRegister.sql
@@ -12,41 +12,31 @@ BEGIN
     SET NOCOUNT ON;
     SET TRANSACTION ISOLATION LEVEL READ COMMITTED;
 
-    BEGIN TRY
-        IF NOT EXISTS (SELECT 1 FROM dbo.Registers WHERE RegisterTemplateId = @RegisterTemplateId AND Subset = @Subset)
-        BEGIN
-            INSERT INTO dbo.Registers (RegisterTemplateId, Subset, Created)
-            VALUES
-            (
-                @RegisterTemplateId
-                ,@Subset
-                ,GETUTCDATE()
-            )
-        END
+    IF NOT EXISTS (SELECT 1 FROM dbo.Registers WHERE RegisterTemplateId = @RegisterTemplateId AND Subset = @Subset)
+    BEGIN
+        INSERT INTO dbo.Registers (RegisterTemplateId, Subset, Created)
+        VALUES
+        (
+            @RegisterTemplateId
+            ,@Subset
+            ,GETUTCDATE()
+        )
+    END
 
-        SELECT TOP 1
-            Id
-            ,RegisterTemplateId
-            ,Subset
-        FROM
-            dbo.Registers
-        WHERE
-            RegisterTemplateId = @RegisterTemplateId
-            AND Subset = @Subset
-
-    END TRY
-    BEGIN CATCH
-        DECLARE @ErrorMessage NVARCHAR(MAX);
-        DECLARE @ErrorSeverity INT;
-        DECLARE @ErrorState INT;
- 
-        SELECT
-            @ErrorMessage = ERROR_MESSAGE(),
-            @ErrorSeverity = ERROR_SEVERITY(),
-            @ErrorState = ERROR_STATE();
- 
-        RAISERROR(@ErrorMessage, @ErrorSeverity, @ErrorState);
-    END CATCH;
+    SELECT TOP 1
+        Id
+        ,RegisterTemplateId
+        ,Subset
+    FROM
+        dbo.Registers
+    WHERE
+        RegisterTemplateId = @RegisterTemplateId
+        AND Subset = @Subset
 
 END
+GO
+
+GRANT EXECUTE
+    ON OBJECT::[dbo].[GetOrCreateRegister] TO [db_execute_procedure_role]
+    AS [dbo];
 GO
