+++ src/backend/Database/Ts.TelemetryDatabase.Sql/Stored Procedures/RegisterSet/UpsertRegisterSet.sql
@@ -4,7 +4,7 @@
 
 CREATE PROCEDURE dbo.GetOrCreateRegisterSet
 (
-    @DeviceId INT,
+    @DeviceId SMALLINT,
     @Identifier NVARCHAR(450)
 )
 AS
@@ -12,41 +12,31 @@ BEGIN
     SET NOCOUNT ON;
     SET TRANSACTION ISOLATION LEVEL READ COMMITTED;
 
-    BEGIN TRY  
-        IF NOT EXISTS (SELECT 1 FROM dbo.RegisterSets WHERE DeviceId = @DeviceId AND Identifier = @Identifier) 
-        BEGIN
-            INSERT INTO dbo.RegisterSets (DeviceId, Identifier, Created)
-            VALUES
-            (
-                @DeviceId
-                ,@Identifier
-                ,GETUTCDATE()
-            )
-        END
+    IF NOT EXISTS (SELECT 1 FROM dbo.RegisterSets WHERE DeviceId = @DeviceId AND Identifier = @Identifier) 
+    BEGIN
+        INSERT INTO dbo.RegisterSets (DeviceId, Identifier, Created)
+        VALUES
+        (
+            @DeviceId
+            ,@Identifier
+            ,GETUTCDATE()
+        )
+    END
 
-        SELECT TOP 1
-            Id
-            ,DeviceId
-            ,Identifier
-        FROM
-            dbo.RegisterSets
-        WHERE
-            DeviceId = @DeviceId
-            AND Identifier = @Identifier
-
-    END TRY
-    BEGIN CATCH
-        DECLARE @ErrorMessage NVARCHAR(MAX);
-        DECLARE @ErrorSeverity INT;
-        DECLARE @ErrorState INT;
- 
-        SELECT
-            @ErrorMessage = ERROR_MESSAGE(),
-            @ErrorSeverity = ERROR_SEVERITY(),
-            @ErrorState = ERROR_STATE();
- 
-        RAISERROR(@ErrorMessage, @ErrorSeverity, @ErrorState);
-    END CATCH;
+    SELECT TOP 1
+        Id
+        ,DeviceId
+        ,Identifier
+    FROM
+        dbo.RegisterSets
+    WHERE
+        DeviceId = @DeviceId
+        AND Identifier = @Identifier
 
 END
+GO
+
+GRANT EXECUTE
+    ON OBJECT::[dbo].[GetOrCreateRegisterSet] TO [db_execute_procedure_role]
+    AS [dbo];
 GO
