+++ src/backend/Database/Ts.TelemetryDatabase.Sql/Stored Procedures/RegisterTemplate/UpsertRegisterTemplate.sql
@@ -16,58 +16,48 @@ BEGIN
     SET NOCOUNT ON;
     SET TRANSACTION ISOLATION LEVEL READ COMMITTED;
 
-    BEGIN TRY  
-        IF NOT EXISTS (SELECT 1 FROM dbo.RegisterTemplates WHERE RegisterSetId = @RegisterSetId AND Identifier = @Identifier)
-        BEGIN
-            INSERT INTO dbo.RegisterTemplates
-            (
-                RegisterSetId
-                ,Identifier
-                ,Name
-                ,Type
-                ,Unit
-                ,Description
-                ,Created
-            )
-            VALUES
-            (
-                @RegisterSetId
-                ,@Identifier
-                ,@Name
-                ,@Type
-                ,@Unit
-                ,@Description
-                ,GETUTCDATE()
-            )
-        END
-
-        SELECT TOP 1
-            Id
-            ,RegisterSetId
+    IF NOT EXISTS (SELECT 1 FROM dbo.RegisterTemplates WHERE RegisterSetId = @RegisterSetId AND Identifier = @Identifier)
+    BEGIN
+        INSERT INTO dbo.RegisterTemplates
+        (
+            RegisterSetId
             ,Identifier
             ,Name
             ,Type
             ,Unit
             ,Description
-        FROM
-            dbo.RegisterTemplates
-        WHERE
-            RegisterSetId = @RegisterSetId
-            AND Identifier = @Identifier
+            ,Created
+        )
+        VALUES
+        (
+            @RegisterSetId
+            ,@Identifier
+            ,@Name
+            ,@Type
+            ,@Unit
+            ,@Description
+            ,GETUTCDATE()
+        )
+    END
 
-    END TRY
-    BEGIN CATCH
-        DECLARE @ErrorMessage NVARCHAR(MAX);
-        DECLARE @ErrorSeverity INT;
-        DECLARE @ErrorState INT;
- 
-        SELECT
-            @ErrorMessage = ERROR_MESSAGE(),
-            @ErrorSeverity = ERROR_SEVERITY(),
-            @ErrorState = ERROR_STATE();
- 
-        RAISERROR(@ErrorMessage, @ErrorSeverity, @ErrorState);
-    END CATCH;
+    SELECT TOP 1
+        Id
+        ,RegisterSetId
+        ,Identifier
+        ,Name
+        ,Type
+        ,Unit
+        ,Description
+    FROM
+        dbo.RegisterTemplates
+    WHERE
+        RegisterSetId = @RegisterSetId
+        AND Identifier = @Identifier
 
 END
+GO
+
+GRANT EXECUTE
+    ON OBJECT::[dbo].[GetOrCreateRegisterTemplate] TO [db_execute_procedure_role]
+    AS [dbo];
 GO
