+++ src/backend/Database/Ts.TelemetryDatabase/Ts.TelemetryDatabase.csproj
@@ -16,4 +16,8 @@
     <PackageReference Include="System.Text.RegularExpressions" />
   </ItemGroup>
 
+  <ItemGroup>
+    <InternalsVisibleTo Include="Ts.Database.Tests" />
+  </ItemGroup>
+
 </Project>
